+++ src/models/project.ts
@@ -1,5 +1,6 @@
 import type { GeometryObject } from './geometryObject'
 import type { Emission } from './material'
+import type { Results } from './result'
 
 /**
  * Project interface, stores all geometry and metadata of the project.
@@ -14,15 +15,4 @@ export interface Project {
 	source?: string
 	location?: string
 	results?: Results[]
-}
-
-/**
- * Results are stored as Impact Category, Life Cycle Stage then emission value
- * ID and date is just for documentation.
- * TODO: Move to results model
- */
-export interface Results {
-	id: string // Run ID for results.
-	date: Date
-	emission: Emission
 }
